requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-filter
psycopg[binary]
python-slugify
pytz
